Draft Fall 2014 - Jutta's comments
This is very well done, I just have a few comments:
1. the framing context, including the two pieces of legislation listed,
is very US centric, do you want to reference some of the other
legislation from around the globe that supports these two "ramps"?
2. Structural navigation is a major accessibility requirement for
anyone not using a mouse or who can't see the visual indicators of
structure (H1 to h1, down to H2, etc.). Could this be added to the
first step especially as the summary (but not the header) is titled
"Text and Solid Structure"? You do allude to the navigation part in
step 2 and the separation of structure and presentation in step 5.
3. I would also add use an ATAG supportive authoring tool to the
template "Things to do" at the end.